making NativeHandle a ulong, rather than IntPtr (#223)

+++ src/Hyperlight/Wrapper/Handle.cs
@@ -1,4 +1,4 @@
-global using NativeHandle = System.IntPtr;
+global using NativeHandle = System.UInt64;
 using System;
 using System.Runtime.InteropServices;
 using Hyperlight.Core;
